fix(test.sh) Force a single nginx worker on tests
Without this setting, the tests are flacky. Usually the problem
manifests when attempting to use an API right after creating it in kong.
If a worker picks the create request and a different worker picks the
proxy request, there's a small time frame in which the API will still not be
"refreshed" on its router.

All kudos to @hishamhm for being able to figure this out via sheer
thought, as he was not able to reproduce the problem on his machine.