feat(test.sh) add new command type: cql
The tests were able to run raw `psql` commands but not cql ones. This commit fixes that. A couple notes:

lua-cassandra throws the following warning:
```
found contact point with '0.0.0.0' as rpc_address, using 'db' to contact it instead. If this is incorrect you should avoid using '0.0.0.0' in rpc_address
```
This doesn't seem to have any impact on the tests, but I could not find a way to make that message dissapear.

Also, if you just look at the diff you will see that `cql` seems to have "replaced" `psql` instead of "adding it in":

```
- if request[1] == "psql" then
-     return parse_command_psql_request(request, context)
+ if request[1] == "cql" then
+    return parse_command_cql_request(request, context)
```

That is not the case. What happened was that previously the `psql` conditional was executed twice. This change replaces the second one with cql.