feat(2.2.0) upgrade paths from 2.1.3 to 2.2.0
These tests execute the non-purely structural part¹ of the migration from 2.1.x to 2.2.0.

In this case, we are testing the target history is erased, and only the latest version of each target (independently of their weight, even when it is 0) remains in the database.

In order to test this in PostgreSQL use the following command:

```
./test.sh -b kong:2.1.4 -t kong:2.2.0-beta.1 --force-migrating upgrade_paths/2.1.3_2.2.0

```

The `--force-migrating` command is needed because we're dealing the the beta version, which the script does not understand. Once 2.2.0 is officially out that part can be removed.

In order to test with cassandra, this is the command:

```
./test.sh -b kong:2.1.4 -t kong:2.2.0-beta.1 --force-migrating -d cassandra upgrade_paths/2.1.3_2.2.0
```

¹ Testing that the necessary tables are created is implicitly tested in Kong's "regular" specs.