chore(ci) fix env vars in old versions of tests, deactivate 1.3.0
For an unknown reason gojira is not able to properly download Kong 1.3.0
any more on these tests, we get an unhelpful error from docker-compose
stating that "the image does not exist any more". So those have been
deactivated.

Using the release/2.2.0 branch now because we found a problem in one of
the cassandra migrations.