tests(2.0.x) add workspaces tests
This introduces a very simple test case (proxying via a route and
service with keyauth) which should be enough to test the workspaces
upgrade.

This is because the Plugins entity has a composite cache key, which is
the one being modified by the migrations.

When the migration works, the composite cache key should be correct at
the end, and the new kong nodes should be able to proxy traffic on the
"migrating" and "after" states.

Notice that in the case of Cassandra, there are a couple exceptions:

* The "migrating" phase is expected to fail, because in Cassandra
  the entity workspaces are updated on the "teardown" phase of
  migrations
* For the same reason the new node needs to be restarted before running
  the `after` migration tests; a new node started before they run
  will not be able to work properly. The new `--target-restart` option
  has been added to enable this restarting.