fix(test.sh) convince `make dev` that go is installed
`make dev` attempts to install grpccurl, which needs go, which isn't installed in most recent images of Kong.

grpccurl is not used by upgrade tests at all, so we can skip this step.

The way we do it is by creating an executable called `go` which always succeeds. The `grpcurl` installation steps will just download grpcurl and then do nothing with it (but won't fail)